20220404 Zigbee2MQTT - old-menu branch - PR 2 of 3
Deprecates Dockerfile-based build in favour of environment variables
that implement the same behaviour.

Dockerfile retained to avoid introducing a breaking change. A
notification is displayed each time the Dockerfile is run:

```
*** DEPRECATION NOTICE: Please read IOTstack Zigbee2MQTT documentation:
*** https://sensorsiot.github.io/IOTstack/Containers/Zigbee2MQTT/
```

The intention is that user attention will be drawn to the need to update
their service definitions.

The revised service definition:

* includes a `depends_on` clause tying Zigbee2MQTT to Mosquitto (the
default arrangement for IOTstack).
* reduces the `devices` list to just `- /dev/ttyAMA0:/dev/ttyACM0`
in favour of extended "how to" documentation for device discovery.

Addresses issues raised in #402, #423 and #538.

Documentation reduced to a stub pointing to master branch.

Signed-off-by: Phill Kelley <34226495+Paraphraser@users.noreply.github.com>